backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
pydantic[email]
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
alembic
email-validator
